was accessing currentUser.uid before there was a user, fixed it.
